requirements.txt
albumentations==1.1.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
imageio==2.13.5
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
munch==2.5.0
networkx==2.6.3
numpy==1.21.5
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==8.4.0
pretrainedmodels==0.7.4
pyparsing==3.0.6
PyWavelets==1.2.0
PyYAML==6.0
qudida==0.0.4
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
tifffile==2021.11.2
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.62.3
typing-extensions==4.0.1
Werkzeug==2.0.2
